+++ src/common/enums/alkemio.error.status.ts
@@ -69,4 +69,5 @@ export enum AlkemioErrorStatus {
   OPERATION_NOT_ALLOWED = 'OPERATION_NOT_ALLOWED',
   NOT_FOUND = 'NOT_FOUND',
   IPFS_NOT_FOUND = 'IPFS_NOT_FOUND',
+  EXCALIDRAW_AMQP_RESULT_ERROR = 'EXCALIDRAW_AMQP_RESULT_ERROR',
 }
