+++ src/services/external/excalidraw-backend/amqp-adapter/index.ts
@@ -1,0 +1,1 @@
+export * from './excalidraw.amqp.server.factory.provider';
