+++ src/services/external/excalidraw-backend/excalidraw.server.module.ts
@@ -4,9 +4,8 @@ import { AuthenticationModule } from '@core/authentication/authentication.module
 import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { ExcalidrawEventPublisherModule } from '@services/excalidraw-pubsub/publisher';
 import { ExcalidrawEventSubscriberModule } from '@services/excalidraw-pubsub/subscriber';
-// import { ExcalidrawAmqpServerFactoryProvider } from './excalidraw.amqp.server.factory.provider';
-import { ExcalidrawRedisServerFactoryProvider } from './excalidraw.redis.server.factory.provider';
 import { APP_ID_PROVIDER } from '@common/app.id.provider';
+import { ExcalidrawRedisServerFactoryProvider } from './redis-adapter';
 
 @Module({
   imports: [
@@ -16,10 +15,6 @@ import { APP_ID_PROVIDER } from '@common/app.id.provider';
     ExcalidrawEventPublisherModule,
     ExcalidrawEventSubscriberModule,
   ],
-  providers: [
-    /*ExcalidrawAmqpServerFactoryProvider, */
-    ExcalidrawRedisServerFactoryProvider,
-    APP_ID_PROVIDER,
-  ],
+  providers: [ExcalidrawRedisServerFactoryProvider, APP_ID_PROVIDER],
 })
 export class ExcalidrawServerModule {}
