+++ src/services/external/excalidraw-backend/redis-adapter/index.ts
@@ -1,0 +1,1 @@
+export * from './excalidraw.redis.server.factory.provider';
